flask/requirements.txt
Flask==1.0.2
flask-cors==3.0.6
redis==2.10.6
numpy==1.15.2
pandas==0.23.4
scipy==1.1.0
scikit-learn==0.20.0